+++ mycodo/mycodo_flask/utils/utils_general.py
@@ -333,7 +333,10 @@ def custom_channel_options_return_json(
 
             if request_form:
                 for key in request_form.keys():
-                    if "{}_{}_{}".format(device_id, channel, each_option['id']) == key:
+                    key_str = "{}_{}_{}".format(device_id, channel, each_option['id'])
+                    if key_str == key:
+                        # logger.error("key_str: {}".format(key_str))
+                        # logger.error("val: {}".format(request_form.get(key)))
                         constraints_pass = True
                         constraints_errors = []
                         value = None